+++ examples/shared/xy-stock-comparing/index.ts
@@ -88,7 +88,7 @@ dateAxisRenderer.labels.template.setAll({
   maxPosition: 0.99
 });
 
-const dateAxis = chart.xAxes.push(am5xy.DateAxis.new(root, {
+const dateAxis = chart.xAxes.push(am5xy.GaplessDateAxis.new(root, {
   groupData: true,
   baseInterval: {
     timeUnit: "day",